a much better fix for the doubleshootyourself bug

+++ code/modules/projectiles/gun.dm
@@ -82,8 +82,6 @@
 	var/gibs = FALSE
 	var/crushes = FALSE
 
-	var/next_shoot_owner = -1
-
 /obj/item/weapon/gun/New()
 	..()
 	if (!firemodes.len)
@@ -437,7 +435,7 @@
 		mouthshoot = FALSE
 		return
 	var/obj/item/projectile/in_chamber = consume_next_projectile()
-	if (istype(in_chamber))
+	if (in_chamber && istype(in_chamber))
 		user.visible_message("<span class = 'warning'>[user] pulls the trigger.</span>")
 		if (silenced)
 			playsound(user, fire_sound, 10, TRUE)
@@ -461,7 +459,15 @@
 		else
 			user << "<span class = 'notice'>Ow...</span>"
 			user.apply_effect(110,AGONY,0)
-		qdel(in_chamber)
+
+		if (istype(src, /obj/item/weapon/gun/projectile))
+			var/obj/item/weapon/gun/projectile/proj = src
+			if (proj.chambered)
+				proj.chambered.expend()
+				proj.process_chambered()
+		else
+			qdel(in_chamber)
+
 		mouthshoot = FALSE
 		return
 	else
@@ -474,8 +480,6 @@
 		return
 	if (!special_check(user))
 		return
-	if (world.time < next_shoot_owner)
-		return
 
 	var/_burst = 1
 	var/datum/firemode/firemode = firemodes[sel_mode]
@@ -486,7 +490,7 @@
 
 		var/obj/item/projectile/in_chamber = consume_next_projectile()
 
-		if (istype(in_chamber))
+		if (in_chamber && istype(in_chamber))
 
 			var/damage_multiplier = 1.33 // so legs and arms don't explode - Kachnov
 			var/organ_name = replacetext(replacetext(user.targeted_organ, "l_", "left "), "r_", "right ")
@@ -511,11 +515,19 @@
 			else
 				user << "<span class = 'notice'>Ow...</span>"
 				user.apply_effect(110,AGONY,0)
-			qdel(in_chamber)
-			next_shoot_owner = world.time + 5 // fixes rare bug where you can shoot yourself multiple times with one bullet
+
+
+			if (istype(src, /obj/item/weapon/gun/projectile))
+				var/obj/item/weapon/gun/projectile/proj = src
+				if (proj.chambered)
+					proj.chambered.expend()
+					proj.process_chambered()
+			else
+				qdel(in_chamber)
+
 		else
 			handle_click_empty(user)
-			return
+			break
 
 /obj/item/weapon/gun/examine(mob/user)
 	..()
